Merge branch 5c44506b 'master' into DA_GFSv16

+++ util/Radiance_Monitor/image_gen/ush/mk_angle_plots.sh
@@ -109,11 +109,6 @@ for type in ${SATYPE}; do
    fi
    ${IG_SCRIPTS}/update_ctl_tdef.sh ${imgndir}/${type}.ctl ${START_DATE} ${NUM_CYCLES}
 
-#   if [[ $MY_MACHINE = "wcoss" || $MY_MACHINE = "zeus" || $MY_MACHINE = "theia" ]]; then
-#      sed -e 's/cray_32bit_ieee/ /' ${imgndir}/${type}.ctl > tmp_${type}.ctl
-#      mv -f tmp_${type}.ctl ${imgndir}/${type}.ctl
-#   fi
-
 done
 
 
@@ -147,7 +142,6 @@ cd $PLOT_WORK_DIR
 
 list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos sin emiss ordang4 ordang3 ordang2 ordang1"
 
-  if [[ ${MY_MACHINE} = "wcoss" || ${MY_MACHINE} = "cray" ]]; then
      suffix=a
      cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
      jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
@@ -158,9 +152,20 @@ list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos 
 
      rm $LOGdir/plot_angle_${suffix}.log
 
+     #--------------------------------------------------
+     # sbatch (slurm) requires a line number added
+     # to the cmdfile
+     ctr=0
      for type in ${SATLIST}; do
-       echo "$IG_SCRIPTS/plot_angle.sh $type $suffix '$list'" >> $cmdfile
+       if [[ ${MY_MACHINE} = "theia" ]]; then
+          echo "${ctr} $IG_SCRIPTS/plot_angle.sh $type $suffix '$list'" >> $cmdfile
+
+       else
+          echo "$IG_SCRIPTS/plot_angle.sh $type $suffix '$list'" >> $cmdfile
+       fi
+       ((ctr=ctr+1))
      done
+
      chmod 755 $cmdfile
      echo "CMDFILE:  $cmdfile"
 
@@ -172,45 +177,27 @@ list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos 
         wall_tm="1:45"
      fi
 
-     if [[ ${MY_MACHINE} = "wcoss" ]]; then
+     if [[ ${MY_MACHINE} = "wcoss" || ${MY_MACHINE} = "wcoss_d" ]]; then
         $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M 20000 -W ${wall_tm} \
              -R affinity[core] -J ${jobname} -cwd ${PWD} $cmdfile
+
+     elif [[ ${MY_MACHINE} = "theia" ]]; then
+        $SUB --account ${ACCOUNT} -n $ctr  -o ${logfile} -D . -J ${jobname} --time=2:00:00 \
+        --wrap "srun -l --multi-prog ${cmdfile}"
+
      else	# cray
         $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M 600 -W ${wall_tm} \
              -J ${jobname} -cwd ${PWD} $cmdfile
      fi
-  else				# Zeus/theia platform
-     for sat in ${SATLIST}; do
-        suffix=${sat} 
-        cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
-        jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
-        logfile=${LOGdir}/plot_angle_${suffix}.log
-
-        rm -f $cmdfile
-        rm -f $logfile
-
-        echo "$IG_SCRIPTS/plot_angle.sh $sat $suffix '$list'" >> $cmdfile
-
-        if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-           wall_tm="5:00:00"
-        else
-           wall_tm="2:30:00"
-        fi
-
-        $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} \
-             -V -j oe -o ${logfile} ${cmdfile}
-     done
-  fi
 
 
 
-  #----------------------------------------------------------------------------
-  #  bigSATLIST
-  #   
-  #    There is so much data for some sat/instrument sources that a separate 
-  #    job for each is necessary.
-  #   
-
+#----------------------------------------------------------------------------
+#  bigSATLIST
+#   
+#    There is so much data for some sat/instrument sources that a separate 
+#    job for each is necessary.
+#   
 echo "starting $bigSATLIST"
 
 set -A list count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos sin emiss ordang4 ordang3 ordang2 ordang1
@@ -218,12 +205,11 @@ set -A list count penalty omgnbc total omgbc fixang lapse lapse2 const scangl cl
 for sat in ${bigSATLIST}; do
    echo processing $sat in $bigSATLIST
 
-   #
+   #--------------------------------------------
    #  wcoss submit 4 jobs for each $sat
    #
-   if [[ $MY_MACHINE = "wcoss" || $MY_MACHINE = "cray" ]]; then 	
+   if [[ $MY_MACHINE = "wcoss" || ${MY_MACHINE} = "wcoss_d" || $MY_MACHINE = "cray" ]]; then 	
       batch=1
-      ii=0
 
       suffix="${sat}_${batch}"
       cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
@@ -231,11 +217,15 @@ for sat in ${bigSATLIST}; do
       jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
       logfile=${LOGdir}/plot_angle_${suffix}.log
 
+      ii=0
       while [[ $ii -le ${#list[@]}-1 ]]; do
 
-         echo "$IG_SCRIPTS/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
-         ntasks=`cat $cmdfile|wc -l `
+         echo "new line in cmdfile = $line"
+         echo $line >> $cmdfile
          chmod 755 $cmdfile
+
+         ntasks=`cat $cmdfile|wc -l `
+         echo "ntasks = $ntasks"
 
          if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
             wall_tm="3:00"
@@ -244,12 +234,9 @@ for sat in ${bigSATLIST}; do
          fi
 
         
-#         mem="6000"
-#         if [[ $batch -eq 1 ]]; then
-            mem="24000"
-#         fi
 
          if [[ $MY_MACHINE = "wcoss" ]]; then
+            mem="24000"
             $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M ${mem} -W ${wall_tm} \
                  -R affinity[core] -J ${jobname} -cwd ${PWD} $cmdfile
          else
@@ -268,31 +255,27 @@ for sat in ${bigSATLIST}; do
          (( ii=ii+1 ))
       done
 
-   else					# Zeus, submit 1 job for each sat/list item
+
+   elif [[ $MY_MACHINE = "theia" ]]; then		# theia, submit 1 job for each sat/list item
 
       ii=0
       suffix="${sat}"
+      logfile=${LOGdir}/plot_angle_${suffix}.log
+      cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
+      rm -f $cmdfile
+
+      logfile=${LOGdir}/plot_angle_${suffix}.log
+      jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
 
       while [[ $ii -le ${#list[@]}-1 ]]; do
-         cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}_${list[$ii]}
-         rm -f $cmdfile
-         logfile=${LOGdir}/plot_angle_${suffix}_${list[$ii]}.log
-         jobname=plot_${RADMON_SUFFIX}_ang_${suffix}_${list[$ii]}
-
-         echo "${IG_SCRIPTS}/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
-
-         if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-            wall_tm="5:00:00"
-         else
-            wall_tm="2:30:00"
-         fi
-
-         $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} \
-              -V -j oe -o ${logfile} ${cmdfile}
-
+         echo "${ii} ${IG_SCRIPTS}/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
          (( ii=ii+1 ))
       done
-  fi
+
+      $SUB --account ${ACCOUNT} -n $ii  -o ${logfile} -D . -J ${jobname} --time=2:00:00 \
+           --wrap "srun -l --multi-prog ${cmdfile}"
+
+   fi
 
 done
